The line:
  nsegdfrm(nbl)      = nsegdfrm(nbl-1)

added to the aeroelastic input section of the code. The purpose is
so that this parameter gets set to a nonzero value for coarser
blocks.  Previously when explicitly entering start and end indices
of an aeroelastic surface and computing only on a coarser mesh, this
parameter was not set to 1 for that coarser level.  This fixes that
coding error.